productionize events
